runtime api working w/ make start and make e2e tests
