+++ pallets/capacity/src/types.rs
@@ -547,24 +547,3 @@ pub trait ProviderBoostRewardsProvider<T: Config> {
 	fn capacity_boost(amount: BalanceOf<T>) -> BalanceOf<T>;
 }
 
-/// Result of checking a Boost History item to see if it's eligible for a reward.
-#[derive(Copy, Clone, Default, Encode, Eq, Decode, RuntimeDebug, MaxEncodedLen, PartialEq, TypeInfo)]
-#[scale_info(skip_type_params(T))]
-pub struct UnclaimedRewardInfo<Balance, BlockNumber >
-where
-	Balance: AtLeast32BitUnsigned,
-	BlockNumber: AtLeast32BitUnsigned,
-{
-	/// The Reward Era for which this reward was earned
-	pub reward_era: RewardEra,
-	/// When this reward expires, i.e. can no longer be claimed
-	pub expires_at_block: BlockNumber,
-	/// The total staked in this era as of the current block
-	pub staked_amount: Balance,
-	/// The amount staked in this era that is eligible for rewards.  Does not count additional amounts
-	/// staked in this era.
-	pub eligible_amount: Balance,
-	/// The amount in token of the reward (only if it can be calculated using only on chain data)
-	pub earned_amount: Balance,
-}
-
